fix(event-not-trigger): lib/config/documentEventAliases.js -> lib/config/documentEventAliases.ts
